Refactor UserProfileViewModel to filter out referenced recipes and update SharedCollectionDetailView to fetch collection owner details
- Updated UserProfileViewModel to filter out recipes that are only referenced by other recipes, ensuring no duplicates are shown.
- Enhanced SharedCollectionDetailView to load the collection owner's information asynchronously and use it when creating shared recipes.
- Added a new image asset for the app icon.
