+++ Cauldron/CauldronIcon.icon/icon.json
@@ -7,7 +7,7 @@
           "fill-specializations" : [
             {
               "value" : {
-                "solid" : "srgb:1.00000,0.57637,0.00000,1.00000"
+                "solid" : "srgb:1.00000,0.57647,0.00000,1.00000"
               }
             },
             {
