+++ Cauldron/Core/Services/CurrentUserSession.swift
@@ -304,18 +304,6 @@ class CurrentUserSession: ObservableObject {
             logger.warning("Failed to set up connection acceptance notifications: \(error.localizedDescription)")
             // Don't block user flow if subscription fails
         }
-
-        // TODO: Re-enable shared recipe notifications once new architecture is implemented
-        /*
-        // Subscribe to shared recipes
-        do {
-            try await dependencies.cloudKitService.subscribeToSharedRecipes(forUserId: userId)
-            logger.info("Successfully set up shared recipe notifications")
-        } catch {
-            logger.warning("Failed to set up shared recipe notifications: \(error.localizedDescription)")
-            // Don't block user flow if subscription fails
-        }
-        */
     }
     
     /// Update user profile
