+++ Cauldron/Core/Utilities/ImageManager.swift
@@ -18,7 +18,9 @@ actor ImageManager {
     init(cloudKitService: CloudKitService) {
         self.cloudKitService = cloudKitService
 
-        let documentsURL = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first!
+        guard let documentsURL = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first else {
+            fatalError("Unable to access documents directory")
+        }
         self.imageDirectoryURL = documentsURL.appendingPathComponent("RecipeImages", isDirectory: true)
 
         // Ensure directory exists
